requirements.txt
affine
catboost
fiona
geopandas
hdbscan
joblib
matplotlib
numpy
pandas
Pillow
rasterio
scikit-image
scikit-learn
scipy
shapely